synchronize command: Split functionality into lib/
The root.go file should only be concerned with managing the Cobra
command. Separate into packages for the configuration object (cfg/),
the API clients (cli/), and the business logic (lib/).
