+++ cli/clients.go
@@ -1,4 +1,4 @@
-package lib
+package cli
 
 import (
 	"context"
@@ -7,6 +7,7 @@ import (
 
 	"github.com/andygrunwald/go-jira"
 	"github.com/cenkalti/backoff"
+	"github.com/coreos/issue-sync/cfg"
 	"github.com/google/go-github/github"
 	"golang.org/x/oauth2"
 )
@@ -16,7 +17,7 @@ import (
 // of the body. If an error occurs during reading, that error is
 // instead printed and returned. This function closes the body for
 // further reading.
-func GetErrorBody(config Config, res *jira.Response) error {
+func GetErrorBody(config cfg.Config, res *jira.Response) error {
 	log := config.GetLogger()
 	defer res.Body.Close()
 	body, err := ioutil.ReadAll(res.Body)
@@ -37,7 +38,7 @@ func GetErrorBody(config Config, res *jira.Response) error {
 // nil HTTP response and a timeout error are returned.
 //
 // It is nearly identical to MakeJIRARequest, but returns a GitHub API response.
-func MakeGHRequest(config Config, f func() (interface{}, *github.Response, error)) (interface{}, *github.Response, error) {
+func MakeGHRequest(config cfg.Config, f func() (interface{}, *github.Response, error)) (interface{}, *github.Response, error) {
 	var ret interface{}
 	var res *github.Response
 	var err error
@@ -66,7 +67,7 @@ func MakeGHRequest(config Config, f func() (interface{}, *github.Response, error
 // nil HTTP response and a timeout error are returned.
 //
 // It is nearly identical to MakeGHRequest, but returns a JIRA API response.
-func MakeJIRARequest(config Config, f func() (interface{}, *jira.Response, error)) (interface{}, *jira.Response, error) {
+func MakeJIRARequest(config cfg.Config, f func() (interface{}, *jira.Response, error)) (interface{}, *jira.Response, error) {
 	var ret interface{}
 	var res *jira.Response
 	var err error
@@ -87,10 +88,10 @@ func MakeJIRARequest(config Config, f func() (interface{}, *jira.Response, error
 	return ret, res, err
 }
 
-// GetGitHubClient initializes a GitHub API client with an OAuth client for authentication,
+// GetGitHubClient initializes a GitHub API cli with an OAuth cli for authentication,
 // then makes an API request to confirm that the service is running and the auth token
 // is valid.
-func GetGitHubClient(config Config) (*github.Client, error) {
+func GetGitHubClient(config cfg.Config) (*github.Client, error) {
 	log := config.GetLogger()
 
 	ctx := context.Background()
@@ -117,22 +118,22 @@ func GetGitHubClient(config Config) (*github.Client, error) {
 	return client, nil
 }
 
-// GetJIRAClient initializes a JIRA API client, then sets the Basic Auth credentials
+// GetJIRAClient initializes a JIRA API cli, then sets the Basic Auth credentials
 // passed to it. (OAuth token support is planned.)
 //
-// The validity of the client and its authentication are not checked here. One way
-// to check them would be to call config.LoadJIRAConfig() after this function.
-func GetJIRAClient(config Config) (*jira.Client, error) {
+// The validity of the cli and its authentication are not checked here. One way
+// to check them would be to call cfg.LoadJIRAConfig() after this function.
+func GetJIRAClient(config cfg.Config) (*jira.Client, error) {
 	log := config.GetLogger()
 
 	client, err := jira.NewClient(nil, config.GetConfigString("jira-uri"))
 	if err != nil {
-		log.Errorf("Error initializing JIRA client; check your base URI. Error: %v", err)
+		log.Errorf("Error initializing JIRA cli; check your base URI. Error: %v", err)
 		return nil, err
 	}
 
 	client.Authentication.SetBasicAuth(config.GetConfigString("jira-user"), config.GetConfigString("jira-pass"))
 
-	log.Debug("JIRA client initialized")
+	log.Debug("JIRA cli initialized")
 	return client, nil
 }
